Add booking and passenger code

+++ src/main/java/lk/ijse/cmjd113/AirTicketCollector/dto/BookingDTO.java
@@ -16,7 +16,7 @@ public class BookingDTO implements Serializable {
     private String bookingId;
     private LocalDateTime bookingDate;
     private Integer seatCount;
-    private BigDecimal totalAmount;
+    private double totalAmount;
     private BookingStatus status;
     private String flightId;
     private String userId;
